Much nicer maps being generated
the maps generated with the diamond-square method are super random and
chaotic, nos resembling anything real whatsoever. plus there's this
thing where the links cross each other, which sucks and I have not found
a sure way to get rid of that without either strangling the cpu or
making the map super weird. Other map generation algorythms will be
implemented in the future, but for now this will suffice, I guess.